Healthcare Pack: fix all 16 CS0618 IntegerValue warnings
Revit 2024+ deprecated ElementId.IntegerValue in favour of
ElementId.Value (Int64). Mirroring the Phase 166 transition pattern
already used elsewhere in the codebase (e.g. UniformityValidator
uses long key = hostRoom.Id.Value).

Changed files:
  Core/Validation/Healthcare/HealthcareValidatorBase.cs   .IntegerValue → .Value
  Core/Validation/Healthcare/MgasFlowValidator.cs         (int)el.Category.Id.Value pattern
  Core/Validation/Healthcare/AntiLigatureValidator.cs     ToHashSet now stores long
  Core/Validation/Healthcare/EndoscopeTraceValidator.cs   .Value comparison
  Core/Twin/IoTDeviceRegistry.cs                          all 4 uses → .Value
  Core/MedGas/MgasNetwork.cs                              (BuiltInCategory)(int)el.Category.Id.Value
  Core/Adjacency/RoomGraphBuilder.cs                      RoomGraph.Adj/.Rooms now Dictionary<long, …>;
                                                          RoomGraph.Neighbours(long); local a/b are long
  Core/Adjacency/CleanDirtyFlowSolver.cs                  FlowFinding.From/ToRoomId now long;
                                                          ClassOf(long); Queue<(long id, …)>;
                                                          HashSet<long> seen
  Docs/Templates/RdsRenderer.cs                           .Value.ToString() fallback

Build now produces 0 warnings on net8.0-windows for the StingTools
plugin assembly.

+++ StingTools/Core/Adjacency/CleanDirtyFlowSolver.cs
@@ -11,8 +11,8 @@ namespace StingTools.Core.Adjacency
 {
     public class FlowFinding
     {
-        public int FromRoomId;
-        public int ToRoomId;
+        public long FromRoomId;
+        public long ToRoomId;
         public string FromClass;
         public string ToClass;
         public string Severity;
@@ -35,7 +35,7 @@ namespace StingTools.Core.Adjacency
             if (doc == null) return findings;
             var g = RoomGraphBuilder.Build(doc);
 
-            string ClassOf(int rid) => g.Rooms.TryGetValue(rid, out var r) ? GetParam(r, "CLN_ROOM_CLASS_TXT") : "";
+            string ClassOf(long rid) => g.Rooms.TryGetValue(rid, out var r) ? GetParam(r, "CLN_ROOM_CLASS_TXT") : "";
 
             foreach (var (rid, room) in g.Rooms)
             {
@@ -43,9 +43,9 @@ namespace StingTools.Core.Adjacency
                 if (string.IsNullOrEmpty(rc) || !DirtySources.Contains(rc)) continue;
 
                 // BFS up to depth 3.
-                var queue = new Queue<(int id, int depth, bool bufferSeen)>();
+                var queue = new Queue<(long id, int depth, bool bufferSeen)>();
                 queue.Enqueue((rid, 0, false));
-                var seen = new HashSet<int> { rid };
+                var seen = new HashSet<long> { rid };
                 while (queue.Count > 0)
                 {
                     var (id, depth, buffered) = queue.Dequeue();
